Move CRO components into submodule
